requirements/dev.txt
ipython==8.31.0
pytest-django==4.8.0
